backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
psycopg2-binary
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
requests
APScheduler
pytz
cryptography
